src/color-correction/requirements.txt
contourpy==1.2.0
coverage==7.4.1
cycler==0.12.1
fonttools==4.47.2
importlib-resources==6.1.1
kiwisolver==1.4.5
matplotlib==3.8.2
numpy==1.26.3
opencv-python==4.9.0.80
packaging==23.2
pillow==10.2.0
pyparsing==3.1.1
python-dateutil==2.8.2
six==1.16.0
zipp==3.17.0
